pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-gerrit"
version = "0.1.0"
description = "MCP server for Gerrit Code Review"
authors = [{name = "Omer Cayirtepe", email = "cayirtepeomer@gmail.com"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
mcp-gerrit = "transport.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
strict = true 
